Refactor reservation detail page into a shared template with modal popup
Implemented using query parameter `popup_view`; which if true, will format
things as modal and use the new base/popup.html template. If the parameter is
not given or false, the template will use the "standalone" format and return
a full featured page extended from the base.html template
