+++ NEMO/templates/base.html
@@ -87,7 +87,7 @@
 
 {# This division is used to display dialog messages via Bootstrap. #}
 <div class="modal fade" id="dialog" tabindex="-1" role="dialog">
-	<div class="modal-dialog">
+	<div class="modal-dialog modal-lg">
 		<div class="modal-content">
 		</div>
 	</div>
