+++ NEMO/templates/base/popup.html
@@ -1,0 +1,7 @@
+<div class="modal-header">
+    <button type="button" class="close" data-dismiss="modal">&times;</button>
+    <h4 class="modal-title">{% block title %}{% endblock %}</h4>
+</div>
+<div class="modal-body">
+    {% block content %}{% endblock %}
+</div>
